perf: :zap: Refactoring to improve performances
