+++ page/src/app.context.js
@@ -1,0 +1,7 @@
+import {createContext, useContext} from 'react';
+
+const CustomContext = createContext();
+
+export const useCustomContext = () => useContext(CustomContext);
+
+export default CustomContext;
