+++ page/src/app.reducer.js
@@ -1,0 +1,10 @@
+const reducer = (state = {events: [], selectedDate: null}, action) => {
+  switch (action.type) {
+    case 'define':
+      return action.payload;
+    default:
+      return state;
+  }
+};
+
+export default reducer;
