+++ page/src/components/Day/Day.utils.js
@@ -1,0 +1,1 @@
+export const displayDate = (date, events) => {};
