+++ page/src/components/Week/Week.js
@@ -1,0 +1,5 @@
+const Week = ({children}) => {
+  return <div className="week">{children}</div>;
+};
+
+export default Week;
