+++ page/src/styles/CalendarGrid.css
@@ -3,6 +3,7 @@
   gap: 1rem;
   flex-wrap: wrap;
   justify-content: center;
+  margin-bottom: 4rem;
 }
 
 .calendarGrid > * {
